requirements.txt
customtkinter==5.2.2
Flask==3.0.3
flask_oauthlib==0.9.6
flask_restful==0.3.10
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==2.2.2
